feat - add numeros com ponto flutuante

+++ script.js
@@ -27,7 +27,6 @@ function onlynumber(evt) {
    var theEvent = evt || window.event;
    var key = theEvent.keyCode || theEvent.which;
    key = String.fromCharCode(key);
-   //var regex = /^[0-9.,]+$/;
    var regex = /^[0-9.]+$/;
    if(!regex.test(key)) {
       theEvent.returnValue = false;
@@ -138,16 +137,16 @@ function storeNumber(number, button) {
 function calcResult(firstNumber, secondNumber) {
     switch (operator) {
         case '+':
-            firstNumber = parseInt(firstNumber) + parseInt(secondNumber);
+            firstNumber = parseFloat(firstNumber) + parseFloat(secondNumber);
             break;
         case '-':
-            firstNumber = parseInt(firstNumber) - parseInt(secondNumber);
+            firstNumber = parseFloat(firstNumber) - parseFloat(secondNumber);
             break;
         case '*':
-            firstNumber = parseInt(firstNumber) * parseInt(secondNumber);
+            firstNumber = parseFloat(firstNumber) * parseFloat(secondNumber);
             break;
         case '/':
-            firstNumber = parseInt(firstNumber) / parseInt(secondNumber);
+            firstNumber = parseFloat(firstNumber) / parseFloat(secondNumber);
             break;
     }   
     return firstNumber;
